Fix mistake when referencing int64 in mathemathics

+++ avutil/mathematics.go
@@ -9,11 +9,11 @@ import "C"
 //
 // C-Function: av_rescale
 func Rescale(a, b, c int64) int64 {
-	return int64(C.av_rescale(C.int64(a), C.int64(b), C.int64(c)))
+	return int64(C.av_rescale(a, b, c))
 }
 
 //
 // C-Function: av_rescale_q
 func RescaleQ(a int64, bq Rational, cq Rational) int64 {
-	return int64(C.av_rescale_q(C.int64(a), (C.struct_AVRational)(bq), (C.struct_AVRational)(cq)))
+	return int64(C.av_rescale_q(a, (C.struct_AVRational)(bq), (C.struct_AVRational)(cq)))
 }
